Add GNU/screen support to debian-installer
GNU/screen doesn't work for all people for sure, it does work for a few
install cases, such as netboot (on a serial console), and
network-console (via SSH).

I surely don't have all devices that d-i supports, and it's confirmed to
be working well on armel/marvell devices only, so I decide to keep
current MEDIUM targets, and to add a few cloned ones with slightly
modification to include screen-udeb.

Here are a few things worth to be mentioned:
 - Created new MEDIUM "netboot-screen", cloned from netboot. But
   netboot-gtk is excluded from adding screen-udeb.
 - Created new MEDIUM "network-screen", cloned from network-console. But
   qnap devices are removed from armel/orion5x/network-screen.cfg, due
   to image size limit.
 - Use symbolic link if possible when cloning from netboot to
   netboot-screen, or from network-console to network-screen.