Set BACKPORTED_SOURCES to linux and net-retriever.
The former is the whole point of implementing backports support; the
latter is needed to get a net-retriever version that knows about the
extra Packages file to be used at runtime.

Warning: That doesn't ensure packages are actually installed from there,
as net-retriever hasn't been backported yet.

Status: for stretch-backports branch only.

Signed-off-by: Cyril Brulebois <cyril@debamax.com>