Bootstrap p-u testing branch for bookworm.
Cherry-pick the first commit of the kibi/bullseye-proposed-updates
branch (and amend it, see below). The other commits were just merges
from the bullseye branch.

Stop setting mirror/udeb/http/hostname (only) in the GRUB config. This
parameter needs to be specified in BIOS mode and in UEFI mode, so let's
have some symmetry, and let kadit handle the addition for both.